so many changes
been busy, finished up evtmgr/evtmgr_cmd/eff_ultra_hammer, along with various other verification, up to 16.66%

+++ include/evt/evt_badgeshop.h
@@ -29,7 +29,7 @@ void badgeShop_bteresaGeneration(void);
 void badgeShop_bargainGeneration(void);
 s32 badgeShop_set(u8* table, s16 id, s16 val);
 s32 badgeShop_get(u8* table, s16 id);
-s32 badgeShop_add(u8* table, s16 id, s16 val);
+s32 badgeShop_add(u8* table, s32 id, s32 val);
 USERFUNC_DEF(badgeShop_getBargainTable);
 USERFUNC_DEF(evt_badgeShop_throw_inc);
 USERFUNC_DEF(evt_badgeShop_throw_dec);
