+++ include/mario/mario_pouch.h
@@ -501,7 +501,7 @@ s32 pouchEquipCheckBadge(s32 badgeId);
 BOOL pouchEquipCheckBadgeIndex(s32 index);
 BOOL pouchEquipBadgeID(s32 badgeId);
 BOOL pouchUnEquipBadgeID(s32 badgeId);
-s32 pouchArriveBadge(s16 id);
+s32 pouchArriveBadge(s32 id);
 void pouchMajinaiInit(u32 curseCount);
 u8 pouchGetKpaCoin(void);
 void pouchAddKpaCoin(u8 coins);
